+++ UI/src/app/modules/validation-result/components/validationrun-row/validationrun-row.component.ts
@@ -90,6 +90,8 @@ export class ValidationrunRowComponent implements OnInit, OnDestroy {
       status = 'Cancelled';
     } else if (valrun.end_time != null) {
       status = 'ERROR';
+    } else {
+      status = 'Running' + ' ' + `${valrun.progress}%`;
     }
     return status;
   }