+++ validator/tests/auxiliary_functions.py
@@ -27,6 +27,32 @@ def generate_default_validation_hires():
 
     return run
 
+def generate_default_validation_hires():
+    run = ValidationRun()
+    run.start_time = datetime.now(tzlocal())
+    run.save()
+
+    data_c = DatasetConfiguration()
+    data_c.validation = run
+    data_c.dataset = Dataset.objects.get(short_name='CGLS_CSAR_SSM1km')
+    data_c.version = DatasetVersion.objects.get(short_name='CGLS_CSAR_SSM1km_V1_1')
+    data_c.variable = DataVariable.objects.get(short_name='CGLS_SSM')
+    data_c.save()
+
+    ref_c = DatasetConfiguration()
+    ref_c.validation = run
+    ref_c.dataset = Dataset.objects.get(short_name='ISMN')
+    ref_c.version = DatasetVersion.objects.get(short_name='ISMN_V20180712_MINI')
+    ref_c.variable = DataVariable.objects.get(short_name='ISMN_soil_moisture')
+    ref_c.save()
+
+    run.reference_configuration = ref_c
+    run.scaling_ref = ref_c
+    run.save()
+
+    return run
+
+
 def generate_default_validation():
     run = ValidationRun()
     run.start_time = datetime.now(tzlocal())
@@ -88,7 +114,7 @@ def generate_default_validation_triple_coll():
     return run
 
 
-def generate_ismn_nonref_validation():
+def generate_ismn_upscaling_validation():
     """Generate a validation where ISMN is used as non-reference"""
     run = ValidationRun()
     run.start_time = datetime.now(tzlocal())