+++ release-notes.md
@@ -1,3 +1,10 @@
+QA4SM v1.8.2 - Release notes 2022-01-07
+=======================================================
+
+# Improvements
+
+1.  Pytesmo and Numpy versions updated
+
 QA4SM v1.8.1 - Release notes 2021-12-01
 =======================================================